feat: move filesystem utils into main gatsby package
