+++ packages/gatsby/src/utils/files/__tests__/create-file-node-from-buffer.js
