+++ packages/gatsby/src/utils/files/__tests__/create-file-node.js
