+++ packages/gatsby/src/utils/files/__tests__/create-remote-file-node.js
