+++ packages/gatsby/src/utils/files/__tests__/fixtures/file.json
@@ -1,0 +1,1 @@
+{"yo": "dog"}
