+++ packages/gatsby/src/utils/files/__tests__/index.js
