+++ packages/gatsby/src/utils/files/__tests__/utils.js
