+++ packages/gatsby/src/utils/files/create-file-node-from-buffer.js
@@ -7,7 +7,7 @@ const { createFilePath } = require(`./utils`)
 const { createContentDigest } = require(`gatsby-core-utils`)
 const cacheId = hash => `create-file-node-from-buffer-${hash}`
 
-/********************
+/** ******************
  * Type Definitions *
  ********************/
 
@@ -103,7 +103,7 @@ async function processBufferNode({
  */
 const processingCache = {}
 
-/***************
+/** *************
  * Entry Point *
  ***************/
 
@@ -150,6 +150,7 @@ module.exports = ({
   }
 
   if (!buffer) {
+    // eslint-disable-next-line prefer-promise-reject-errors
     return Promise.reject(`bad buffer: ${buffer}`)
   }
