+++ packages/gatsby/src/utils/files/create-file-node.js
