+++ packages/gatsby/src/utils/files/create-file-path.js
