+++ packages/gatsby/utils.js
@@ -1,1 +1,5 @@
 exports.createContentDigest = require(`gatsby-core-utils`).createContentDigest
+exports.createFilePath = require(`./src/utils/files/create-file-path`)
+exports.createRemoteFileNode = require(`./src/utils/files/create-remote-file-node`)
+exports.createFileNodeFromBuffer = require(`./src/utils/files/create-file-node-from-buffer`)
+exports.loadNodeContent = require(`./src/utils/files/utils`)
